Pause and resume methods have been implemented.
There is also a new method to determine if the carousel has an
associated animation: hasActiveAnimation().
The carousel can now pause animations.
There is not easy to wait for a certain amount of time ignoring the
pause time, so we've added the method createWaitPromise for that
purpose.
HIDDEN alias has been renamed to SLIDE_HIDDEN.

We need to add test for the methods pause, resume, createWaitPromise and
hasActiveAnimation